split more code into modules but temporary mostly
