+++ index.html
@@ -7,7 +7,7 @@
   <body>
     <header>
       <div id="title">Binary Tree Visualization</div>
-      <label>Enter an array:</label><br />
+      <label id="array-input-label">Enter an array:</label>
       <input type="text" id="array-input" name="array" aria-label="array-input" value="3,4,2,9,8,7,12"/>
       <ul id="tree-options">
         <li id="binary-tree-unsorted" class="active-option">Binary Tree (Unsorted)</li>
